# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Facility Management System API",
        "description": "API documentation for the Facility Management System",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Local Environment"
        }
    ],
    "paths": {
...
        "/api/maintenance-logs": {
            "get": {
                "tags": [
                    "Maintenance Logs"
                ],
                "summary": "Get all maintenance logs",
                "operationId": "d51fc8f22fead0512e308ebaeb8ded0d",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of maintenance logs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Maintenance logs retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MaintenanceLog"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Maintenance Logs"
                ],
                "summary": "Create a new maintenance log",
                "operationId": "dbcff43fd1ba8577192b4bfa888a9bdc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MaintenanceLog"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Maintenance log created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Maintenance log created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/MaintenanceLog"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/maintenance-logs/{id}": {
            "get": {
                "tags": [
                    "Maintenance Logs"
                ],
                "summary": "Get a maintenance log by ID",
                "operationId": "14ac8dd1c3f46efe8cfff2bb20ccc716",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Maintenance log ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Maintenance log details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Maintenance log retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/MaintenanceLog"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Maintenance Logs"
                ],
                "summary": "Update a maintenance log",
                "operationId": "d4d8dc0d2e360b64aeb32d3d366ef8c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Maintenance log ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MaintenanceLog"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Maintenance log updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Maintenance log updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/MaintenanceLog"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Maintenance Logs"
                ],
                "summary": "Delete a maintenance log",
                "operationId": "178ca29855a0b6de3c263350940638ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Maintenance log ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Maintenance log deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Maintenance log deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/maintenance-logs/asset/{assetId}": {
            "get": {
                "tags": [
                    "Maintenance Logs"
                ],
                "summary": "Get maintenance logs for an asset",
                "operationId": "0b83453910abed2eaf0eaa8197baadc4",
                "parameters": [
                    {
                        "name": "assetId",
                        "in": "path",
                        "description": "Asset ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of maintenance logs for the asset",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Asset maintenance logs retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MaintenanceLog"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/maintenance-logs/schedule/{scheduleId}": {
            "get": {
                "tags": [
                    "Maintenance Logs"
                ],
                "summary": "Get maintenance logs for a schedule",
                "operationId": "8682c5ec722fd2243a35621bbe0cdf8d",
                "parameters": [
                    {
                        "name": "scheduleId",
                        "in": "path",
                        "description": "Schedule ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of maintenance logs for the schedule",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Schedule maintenance logs retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MaintenanceLog"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}